+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Abstractions/Users/IUserRoleService.cs
@@ -1,0 +1,9 @@
+using Command = TimeSheet.Modules.EmploymentManagement.Application.Users.AssignRole.Command;
+using Result = TimeSheet.Modules.EmploymentManagement.Application.Users.AssignRole.Result;
+
+namespace TimeSheet.Modules.EmploymentManagement.Application.Abstractions.Users;
+
+public interface IUserRoleService
+{
+    Task<Result> AssignRoleAsync(Command command, CancellationToken cancellationToken);
+}