+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Users/AssignRole/Command.cs
@@ -1,0 +1,3 @@
+namespace TimeSheet.Modules.EmploymentManagement.Application.Users.AssignRole;
+
+public sealed record Command(Guid UserId, Guid RoleId);