+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Infrastructure/DependencyInjection.cs
@@ -8,6 +8,7 @@ using TimeSheet.Modules.EmploymentManagement.Application.Abstractions.Audit;
 using TimeSheet.Modules.EmploymentManagement.Application.Abstractions.Employees;
 using TimeSheet.Modules.EmploymentManagement.Application.Abstractions.Roles;
 using TimeSheet.Modules.EmploymentManagement.Application.Abstractions.Security;
+using TimeSheet.Modules.EmploymentManagement.Application.Abstractions.Users;
 using TimeSheet.Modules.EmploymentManagement.Application.Authentication.Configuration;
 using TimeSheet.Modules.EmploymentManagement.Infrastructure.Authentication;
 using TimeSheet.Modules.EmploymentManagement.Infrastructure.Auditing;
@@ -15,6 +16,7 @@ using TimeSheet.Modules.EmploymentManagement.Infrastructure.Employees;
 using TimeSheet.Modules.EmploymentManagement.Infrastructure.Initialization;
 using TimeSheet.Modules.EmploymentManagement.Infrastructure.Persistence;
 using TimeSheet.Modules.EmploymentManagement.Infrastructure.Roles;
+using TimeSheet.Modules.EmploymentManagement.Infrastructure.Users;
 
 namespace TimeSheet.Modules.EmploymentManagement.Infrastructure;
 
@@ -50,6 +52,7 @@ public static class DependencyInjection
         services.AddScoped<IEmployeeWriteService, EmployeeWriteService>();
         services.AddScoped<IEmployeeAddressService, EmployeeAddressService>();
         services.AddScoped<IRoleReadService, RoleReadService>();
+        services.AddScoped<IUserRoleService, UserRoleService>();
         services.AddScoped<DatabaseInitializer>();
 
         return services;